refactor naming structure
