+++ src/api/documents/active_site.py
@@ -4,7 +4,7 @@ import datetime
 
 # Third-Party Libraries
 from bson.objectid import ObjectId
-from utils.db_utils import Document, db
+from utils.db import Document, db
 
 
 class ActiveSite(Document):
